+++ CHANGELOG.md
@@ -73,6 +73,7 @@ and this project adheres to
 - retry Resana file download and skip failing files instead of aborting #175
 - exclude Resana workspaces where user isn't Animateur from migration #110 #160
 - exclude Resana personal workspaces from migration by default #163
+- fix user.csv not generated for locked Resana workspaces #169
 
 ### Removed